feat: split signing barrel into client/server sub-barrels (adcp-client#583 item 3)
Additive pulled forward from #583 so #578's auto-wiring imports and any
3.x consumers can adopt the narrower surface now, and #583 can drop the
monolithic barrel cleanly at the next major.

- `@adcp/client/signing/client` — signer, canonicalization helpers, fetch
  wrapper, capability cache, and the auto-wiring helpers a buyer building
  an AdCPClient needs.
- `@adcp/client/signing/server` — verifier pipeline, Express-shaped
  middleware, JWKS / replay / revocation stores, error taxonomy.
- `@adcp/client/signing` now re-exports the union of both for back-compat.
  Coding agents reading a file cold only need to hold one half of the
  taxonomy; existing consumers keep working unchanged.

package.json exports + typesVersions updated; changeset amended.
