+++ src/lib/signing/index.ts
@@ -1,55 +1,15 @@
-export {
-  buildSignatureBase,
-  canonicalAuthority,
-  canonicalMethod,
-  canonicalTargetUri,
-  formatSignatureParams,
-  getHeaderValue,
-  type RequestLike,
-  type SignatureParams,
-} from './canonicalize';
-export { computeContentDigest, contentDigestMatches, parseContentDigest } from './content-digest';
-export { jwkToPublicKey, verifySignature } from './crypto';
-export { RequestSignatureError, type RequestSignatureErrorCode } from './errors';
-export { StaticJwksResolver, type JwksResolver } from './jwks';
-export { parseSignature, parseSignatureInput, type ParsedSignature, type ParsedSignatureInput } from './parser';
-export {
-  InMemoryReplayStore,
-  type InMemoryReplayStoreOptions,
-  type ReplayInsertResult,
-  type ReplayStore,
-} from './replay';
-export { InMemoryRevocationStore, type RevocationStore } from './revocation';
-export {
-  ALLOWED_ALGS,
-  CLOCK_SKEW_TOLERANCE_SECONDS,
-  MANDATORY_COMPONENTS,
-  MAX_SIGNATURE_WINDOW_SECONDS,
-  REQUEST_SIGNING_TAG,
-  type AdcpJsonWebKey,
-  type ContentDigestPolicy,
-  type RevocationSnapshot,
-  type VerifiedSigner,
-  type VerifierCapability,
-} from './types';
-export { verifyRequestSignature, type VerifyRequestOptions } from './verifier';
-export { signRequest, type SignedRequest, type SignerKey, type SignRequestOptions } from './signer';
-export { createSigningFetch, type CoverContentDigestPredicate, type SigningFetchOptions } from './fetch';
-export { createExpressVerifier, type ExpressLike, type ExpressMiddlewareOptions } from './middleware';
-export {
-  CapabilityCache,
-  buildCapabilityCacheKey,
-  defaultCapabilityCache,
-  type CachedCapability,
-  type CapabilityCacheOptions,
-} from './capability-cache';
-export {
-  buildAgentSigningFetch,
-  extractAdcpOperation,
-  resolveCoverContentDigest,
-  shouldSignOperation,
-  toSignerKey,
-  type BuildAgentSigningFetchOptions,
-} from './agent-fetch';
-export { buildAgentSigningContext, type AgentSigningContext } from './agent-context';
-export { ensureCapabilityLoaded, CAPABILITY_OP } from './capability-priming';
+/**
+ * Aggregate signing barrel. Re-exports the `client` and `server` sub-barrels
+ * so existing consumers of `@adcp/client/signing` keep working.
+ *
+ * New code should import from the narrower surface that matches its role:
+ *
+ *   import { createSigningFetch } from '@adcp/client/signing/client';
+ *   import { createExpressVerifier } from '@adcp/client/signing/server';
+ *
+ * — coding agents reading a file cold only need to hold one half of the
+ * taxonomy, not all 30+ symbols. This aggregate barrel remains the stable
+ * entry point until a future major release drops it.
+ */
+export * from './client';
+export * from './server';
